# api/openapi.json
{
  "components": {
    "schemas": {
...
      "FavoritesOut": {
        "additionalProperties": {
          "items": {
            "type": "string"
          },
          "type": "array"
        },
        "description": "The caller's own starred ids, keyed by object type.\n\nOne key per registered favoritable kind, always, and an empty list where\nthere are no stars: a person with no favorites has an empty set, not an\nunknown one, and a client that has to tell those apart will get it wrong.",
        "title": "FavoritesOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "veodyn-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/favorites": {
      "get": {
        "description": "Every id this caller has starred, grouped by kind.\n\nWhole sets rather than a page: the client marks rows by membership, which\nneeds all of them. One key per registered favoritable kind, present even when\nempty, and a stored row whose kind is no longer registered is left out.",
        "operationId": "list_favorites_favorites_get",
        "parameters": [
          {
            "in": "header",
            "name": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cookie"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Favorites",
        "tags": [
          "favorites"
        ]
      }
    },
    "/favorites/{object_type}/{object_id}": {
      "delete": {
        "description": "Remove one star, whether or not the object still exists.\n\nRemoving the star of something already deleted is the tidy-up a caller should\nbe able to finish. The kind is still resolved, so a delete cannot address rows\nof a kind this build does not serve.",
        "operationId": "remove_favorite_favorites__object_type___object_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "object_type",
            "required": true,
            "schema": {
              "title": "Object Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cookie"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Favorite",
        "tags": [
          "favorites"
        ]
      },
      "post": {
        "operationId": "add_favorite_favorites__object_type___object_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "object_type",
            "required": true,
            "schema": {
              "title": "Object Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cookie"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Favorite",
        "tags": [
          "favorites"
        ]
      }
    },
...
  }
}